--- a/.travis/lts-6.0.yaml
+++ b/.travis/lts-6.0.yaml
@@ -10,5 +10,4 @@ packages:
 extra-deps:
 - hedgehog-0.5.2
 - leancheck-0.6.7
-- tasty-hedgehog-0.1.0.1
 - wl-pprint-annotated-0.1.0.0