--- a/concurrency/concurrency.cabal
+++ b/concurrency/concurrency.cabal
@@ -2,7 +2,7 @@
 -- documentation, see http://haskell.org/cabal/users-guide/
 
 name:                concurrency
-version:             1.4.0.2
+version:             1.5.0.0
 synopsis:            Typeclasses, functions, and data types for concurrency and STM.
 
 description:
@@ -32,7 +32,7 @@ source-repository head
 source-repository this
   type:     git
   location: https://github.com/barrucadu/dejafu.git
-  tag:      concurrency-1.4.0.2
+  tag:      concurrency-1.5.0.0
 
 library
   exposed-modules:     Control.Monad.Conc.Class
@@ -55,14 +55,14 @@ library
 
   -- other-modules:       
   -- other-extensions:    
-  build-depends:       base              >=4.8  && <5
-                     , array             >=0.5  && <0.6
+  build-depends:       base              >=4.9  && <5
+                     , array             >=0.5.1 && <0.6
                      , atomic-primops    >=0.8  && <0.9
                      , exceptions        >=0.7  && <0.11
                      , monad-control     >=1.0  && <1.1
                      , mtl               >=2.2  && <2.3
                      , stm               >=2.4  && <2.5
-                     , transformers      >=0.4  && <0.6
+                     , transformers      >=0.5  && <0.6
   -- hs-source-dirs:      
   default-language:    Haskell2010
   ghc-options:         -Wall