--- a/dejafu-tests/dejafu-tests.cabal
+++ b/dejafu-tests/dejafu-tests.cabal
@@ -43,6 +43,7 @@ library
 
                      , Common
                      , QSemN
+                     , Test.Tasty.Hedgehog
 
   build-depends:       base
                      , abstract-deque
@@ -58,7 +59,6 @@ library
                      , tasty
                      , tasty-expected-failure
                      , tasty-dejafu
-                     , tasty-hedgehog
                      , tasty-hunit
                      , vector
   if impl(ghc < 8.0.1)