--- a/dejafu-tests/dejafu-tests.cabal
+++ b/dejafu-tests/dejafu-tests.cabal
@@ -61,8 +61,6 @@ library
                      , tasty-dejafu
                      , tasty-hunit
                      , vector
-  if impl(ghc < 8.0.1)
-    build-depends: transformers
   hs-source-dirs:      lib
   default-language:    Haskell2010
   ghc-options:         -Wall