--- a/dejafu-tests/lib/Common.hs
+++ b/dejafu-tests/lib/Common.hs
@@ -1,3 +1,5 @@
+{-# OPTIONS_GHC -Wno-deprecations -Wno-orphans #-}
+
 {-# LANGUAGE GADTs #-}
 
 module Common (module Common, module Test.Tasty.DejaFu, T.TestTree, T.expectFail) where