--- a/dejafu-tests/lib/Integration/SCT.hs
+++ b/dejafu-tests/lib/Integration/SCT.hs
@@ -1,3 +1,5 @@
+{-# OPTIONS_GHC -Wno-deprecations #-}
+
 module Integration.SCT where
 
 import           Control.Concurrent.Classy hiding (check)