--- a/dejafu/dejafu.cabal
+++ b/dejafu/dejafu.cabal
@@ -2,7 +2,7 @@
 -- documentation, see http://haskell.org/cabal/users-guide/
 
 name:                dejafu
-version:             1.4.0.0
+version:             1.5.0.0
 synopsis:            A library for unit-testing concurrent programs.
 
 description:
@@ -33,7 +33,7 @@ source-repository head
 source-repository this
   type:     git
   location: https://github.com/barrucadu/dejafu.git
-  tag:      dejafu-1.4.0.0
+  tag:      dejafu-1.5.0.0
 
 library
   exposed-modules:     Test.DejaFu
@@ -58,17 +58,15 @@ library
 
   -- other-modules:       
   -- other-extensions:    
-  build-depends:       base              >=4.8 && <5
-                     , concurrency       >=1.3 && <1.5
+  build-depends:       base              >=4.9 && <5
+                     , concurrency       >=1.5 && <1.6
                      , containers        >=0.5 && <0.6
                      , deepseq           >=1.1 && <2
                      , exceptions        >=0.7 && <0.11
                      , leancheck         >=0.6 && <0.8
                      , profunctors       >=4.0 && <6.0
                      , random            >=1.0 && <1.2
-                     , transformers      >=0.4 && <0.6
+                     , transformers      >=0.5 && <0.6
   -- hs-source-dirs:      
   default-language:    Haskell2010
   ghc-options:         -Wall
-  if impl(ghc < 8.0.1)
-    build-depends: semigroups >=0.16 && <0.19