--- a/hunit-dejafu/hunit-dejafu.cabal
+++ b/hunit-dejafu/hunit-dejafu.cabal
@@ -2,7 +2,7 @@
 -- documentation, see http://haskell.org/cabal/users-guide/
 
 name:                hunit-dejafu
-version:             1.1.0.3
+version:             1.2.0.0
 synopsis:            Deja Fu support for the HUnit test framework.
 
 description:
@@ -30,15 +30,15 @@ source-repository head
 source-repository this
   type:     git
   location: https://github.com/barrucadu/dejafu.git
-  tag:      hunit-dejafu-1.1.0.3
+  tag:      hunit-dejafu-1.2.0.0
 
 library
   exposed-modules:     Test.HUnit.DejaFu
   -- other-modules:       
   -- other-extensions:    
-  build-depends:       base       >=4.8 && <5
+  build-depends:       base       >=4.9 && <5
                      , exceptions >=0.7 && <0.11
-                     , dejafu     >=1.2 && <1.5
-                     , HUnit      >=1.2 && <1.7
+                     , dejafu     >=1.5 && <1.6
+                     , HUnit      >=1.3.1 && <1.7
   -- hs-source-dirs:      
   default-language:    Haskell2010