… .weeder.yaml
… concurrency/CHANGELOG.rst
… dejafu-tests/lib/Test/Tasty/Hedgehog.hs
… dejafu/CHANGELOG.rst
… dejafu/Test/DejaFu/Conc/Internal.hs
… dejafu/Test/DejaFu/Conc/Internal/Memory.hs
… dejafu/Test/DejaFu/Conc/Internal/STM.hs
… dejafu/Test/DejaFu/Conc/Internal/Threading.hs
… dejafu/Test/DejaFu/Internal.hs
… dejafu/Test/DejaFu/SCT/Internal.hs
… dejafu/Test/DejaFu/SCT/Internal/DPOR.hs
… doc/getting_started.rst
… doc/ghc.rst
… hunit-dejafu/CHANGELOG.rst
… tasty-dejafu/CHANGELOG.rst