## Summary

Now that GHC 8.4 is out, we can drop GHC 7.10 and implement some stuff which compatibility prevented us from:

- Putting `forkOSWithUnmask` into `MonadConc` and moving `fork`, `forkOn`, `forkOS`, and `forkOSN` out
- Using `HasCallStack` for all internal errors
- Dropping a bunch of CPP

`HasCallStack` errors look like this:

```
Exception: (dejafu) adjust: key 'main' not found
CallStack (from HasCallStack):
  eadjust, called at ./Test/DejaFu/Conc/Internal/Threading.hs:176:11 in dejafu-1.5.0.0-HUqUYeGD0RYj1ji2N49Cg:Test.DejaFu.Conc.Internal.Threading
  makeBound, called at ./Test/DejaFu/Conc/Internal.hs:112:49 in dejafu-1.5.0.0-HUqUYeGD0RYj1ji2N49Cg:Test.DejaFu.Conc.Internal
  runConcurrency', called at ./Test/DejaFu/Conc/Internal.hs:93:10 in dejafu-1.5.0.0-HUqUYeGD0RYj1ji2N49Cg:Test.DejaFu.Conc.Internal
  runConcurrency, called at ./Test/DejaFu/Conc.hs:220:10 in dejafu-1.5.0.0-HUqUYeGD0RYj1ji2N49Cg:Test.DejaFu.Conc

Exception: (dejafu) taken thread does not exist in done set
CallStack (from HasCallStack):
  fatal, called at ./Test/DejaFu/SCT/Internal/DPOR.hs:80:48 in dejafu-1.5.0.0-HUqUYeGD0RYj1ji2N49Cg:Test.DejaFu.SCT.Internal.DPOR
  validateDPOR, called at ./Test/DejaFu/SCT/Internal/DPOR.hs:175:15 in dejafu-1.5.0.0-HUqUYeGD0RYj1ji2N49Cg:Test.DejaFu.SCT.Internal.DPOR
  incorporateTrace, called at ./Test/DejaFu/SCT.hs:207:25 in dejafu-1.5.0.0-HUqUYeGD0RYj1ji2N49Cg:Test.DejaFu.SCT
```

Also drops the dejafu-tests dep on tasty-hedgehog, as inlining a single module to avoid Stackage nightly being broken seems like a good move.

**Related issues:** closes #132, closes #216, and closes #234 

## Benchmark results

**Before:**

```
$ stack exec -- dejafu-tests +RTS -s

 199,532,749,016 bytes allocated in the heap
  52,859,183,392 bytes copied during GC
      13,756,904 bytes maximum residency (14931 sample(s))
         147,904 bytes maximum slop
              38 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     177320 colls,     0 par   100.044s  101.309s     0.0006s    0.0499s
  Gen  1     14931 colls,     0 par    8.278s   8.386s     0.0006s    0.0108s

  TASKS: 237457 (237452 bound, 5 peak workers (5 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time  152.950s  (159.163s elapsed)
  GC      time  108.322s  (109.695s elapsed)
  EXIT    time    0.001s  (  0.002s elapsed)
  Total   time  261.274s  (268.860s elapsed)

  Alloc rate    1,304,558,648 bytes per MUT second

  Productivity  58.5% of total user, 59.2% of total elapsed
```

**After:**

```
$ stack exec -- dejafu-tests +RTS -s

 202,563,813,528 bytes allocated in the heap
  53,387,891,768 bytes copied during GC
      13,954,624 bytes maximum residency (14989 sample(s))
         152,336 bytes maximum slop
              38 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     180324 colls,     0 par   100.209s  101.551s     0.0006s    0.0562s
  Gen  1     14989 colls,     0 par    8.394s   8.490s     0.0006s    0.0091s

  TASKS: 237802 (237797 bound, 5 peak workers (5 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.001s  (  0.001s elapsed)
  MUT     time  162.331s  (169.462s elapsed)
  GC      time  108.603s  (110.042s elapsed)
  EXIT    time    0.001s  (  0.006s elapsed)
  Total   time  270.936s  (279.510s elapsed)

  Alloc rate    1,247,844,121 bytes per MUT second

  Productivity  59.9% of total user, 60.6% of total elapsed
```